Added Pause Check Box
Added a check box to pause a story.  Also, check state and pause status
on template load of edit story.  This works to make sure that the next
state button should only be visible when appropriate for template load.
the change_state is still as before, and needs to be kept separate for
special cases.  Ask if you do not understand why both are still needed.